% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The_question.R
\name{The_question}
\alias{The_question}
\title{Give the answer.}
\usage{
The_question(x)
}
\arguments{
\item{x}{ANYTHING}
}
\value{
The answer.
}
\description{
Give the answer.
}
\examples{
The_question(x=1337)
}
